+++ apis/getambassador.io/v1beta2/filter_oauth2.go
@@ -24,13 +24,14 @@ type FilterOAuth2 struct {
 	GrantType string `json:"grantType"`
 
 	// grantType=AuthorizationCode
-	RawClientURL       string   `json:"clientURL"` // formerly tenant.tenantUrl
-	ClientURL          *url.URL `json:"-"`         // calculated from RawClientURL
-	DeprecatedStateTTL string   `json:"stateTTL"`
-	ClientID           string   `json:"clientID"`
-	Secret             string   `json:"secret"`
-	SecretName         string   `json:"secretName"`
-	SecretNamespace    string   `json:"secretNamespace"`
+	RawClientURL       string             `json:"clientURL"` // formerly tenant.tenantUrl
+	ClientURL          *url.URL           `json:"-"`         // calculated from RawClientURL
+	DeprecatedStateTTL string             `json:"stateTTL"`
+	ClientID           string             `json:"clientID"`
+	Secret             string             `json:"secret"`
+	SecretName         string             `json:"secretName"`
+	SecretNamespace    string             `json:"secretNamespace"`
+	UseSessionCookies  *UseSessionCookies `json:"useSessionCookies"`
 
 	RawMaxStale string        `json:"maxStale"`
 	MaxStale    time.Duration `json:"-"` // calculated from RawMaxStale
@@ -43,6 +44,19 @@ type FilterOAuth2 struct {
 
 	AccessTokenValidation string             `json:"accessTokenValidation"`
 	AccessTokenJWTFilter  JWTFilterReference `json:"accessTokenJWTFilter"`
+}
+
+type UseSessionCookies struct {
+	Value           *bool               `json:"value"`
+	IfRequestHeader HeaderFieldSelector `json:"ifRequestHeader"`
+}
+
+func (m *UseSessionCookies) Validate() error {
+	if m.Value == nil {
+		value := true
+		m.Value = &value
+	}
+	return m.IfRequestHeader.Validate()
 }
 
 type JWTFilterReference struct {
@@ -92,6 +106,16 @@ func (m *FilterOAuth2) Validate(namespace string, secretsGetter coreV1client.Sec
 				return errors.Errorf("secret name=%q namespace=%q does not contain an oauth2-client-secret field", m.SecretName, m.SecretNamespace)
 			}
 			m.Secret = string(secretVal)
+		}
+
+		if m.UseSessionCookies == nil {
+			value := false
+			m.UseSessionCookies = &UseSessionCookies{
+				Value: &value,
+			}
+		}
+		if err := m.UseSessionCookies.Validate(); err != nil {
+			return errors.Wrap(err, "useSessionCookies")
 		}
 	case GrantType_ClientCredentials:
 		if m.RawClientURL != "" {